+++ packages/http-common/README.md
@@ -1,0 +1,4 @@
+
+# @sensejs/http-common
+
+Common facility for HTTP protocol, for both client side and server side
