+++ packages/http-common/tests/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../../tsconfig.test",
+}
