+++ packages/http-common/tsconfig.build.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./tsconfig",
+  "compilerOptions": {
+    "noEmit": false,
+    "outDir": "lib",
+    "baseUrl": ".",
+    "paths": {}
+  }
+}
