+++ packages/http/src/index.ts
@@ -1,4 +1,3 @@
-export * from './http-abstract';
-export * from './http-decorators';
+export * from '@sensejs/http-common';
 export * from './http-koa-integration';
 export * from './http-module';
